lo intente xd

+++ Public/vistas/Docentes/buscar-docentes.js
@@ -11,7 +11,7 @@ export function modulo(){
         $("#txtNombreDocente").value = docente.nombre;
         $("#txtDireccionDocente").value = docente.direccion;
         $("#txtTelefonoDocente").value = docente.telefono;
-        $("#txtEscalafonDocente").value = docente.escalafon;
+    
     };
     let eliminarDocente = (idDocente)=>{
         fetch(`private/Modulos/Docentes/procesosdoc.php?proceso=eliminarDocente&docente=${idDocente}`).then( resp=>resp.json() ).then(resp=>{
@@ -28,7 +28,7 @@ export function modulo(){
                         <td>${docente.nombre}</td>
                         <td>${docente.direccion}</td>
                         <td>${docente.telefono}</td>
-                        <td>${docente.escalafon}</td>
+                      
                         <td>
                             <input type="button" class="btn btn-outline-danger text-white" value="del">
                         </td>
@@ -38,7 +38,11 @@ export function modulo(){
             $("#tbl-buscar-docentes > tbody").innerHTML = filas;
             $("#tbl-buscar-docentes > tbody").addEventListener("click",e=>{
                 if( e.srcElement.parentNode.dataset.docentes==null ){
+                    let confirmacion = confirm(`¿Seguro que quiere eliminar el registro?`)
+
+                    if (confirmacion == true) {
                     eliminarDocente( e.srcElement.parentNode.parentNode.dataset.idDocente );
+                    }
                 } else {
                     modificarDocente( JSON.parse(e.srcElement.parentNode.dataset.docente) );
                 }
